+++ app/Packages/PodcastAIKit/Sources/PodcastAIKit/Exports.swift
@@ -14,6 +14,7 @@
 @_exported import PodcastAIPlayback
 @_exported import PodcastAISmartFeeds
 @_exported import PodcastAIExport
+@_exported import PodcastAIWidgetData
 
 #if canImport(SwiftData)
 @_exported import PodcastAIPersistence